Move integration test into test_stage.
